Add keyboard shortcuts to select sessions
Motivation: I wanted to switch sessions with minimal keystroke. e.g. run
Hyprland for the most, but use awesome if I want to play some games.

By default, it assign F8 to be "Hyprland", and F9 to be "awesome" as a
session config. It will do nothing if session is not found corresponding
with the name.

It can be extended to support configurable session names, simply not
implemented.